remove fields activity update:
choose the disiered area
click on a GYM area that you want to remove
click on remove
click on "Yes"
the users will no longer be able to sign on that GYM area.

choose the disiered field
click on a soccer field that you want to remove
click on remove
click on "Yes"
the users will no longer be able to sign on that soccer field.

choose the disiered field
click on a basketball field that you want to remove
click on remove
click on "Yes"
the users will no longer be able to sign on that basketball field.